chore: unify post and page header controls
